+++ packages/gatsby-source-wordpress/src/__tests__/normalize.js
@@ -56,7 +56,10 @@ describe(`Process WordPress data`, () => {
   })
 
   it(`Removes the acf key when acf is not an object`, () => {
-    let dummyEntities = [{ id: 1, acf: false }, { id: 2, acf: {} }]
+    let dummyEntities = [
+      { id: 1, acf: false },
+      { id: 2, acf: {} },
+    ]
     expect(normalize.normalizeACF(dummyEntities)).toEqual([
       { id: 1 },
       { id: 2, acf: {} },