+++ www/src/assets/icons/layer-icon.js
@@ -1,7 +1,7 @@
 import React from "react"
-import { colors } from "gatsby-design-tokens"
+import { colors } from "gatsby-design-tokens/dist/theme-gatsbyjs-org"
 
-const LayerIcon = ({ name, fillColor = colors.grey[`50`] }) => {
+const LayerIcon = ({ name, fillColor = colors.grey[50] }) => {
   const icons = {
     Content: (
       <svg