+++ www/src/components/features/evaluation-table.js
@@ -1,7 +1,7 @@
 /** @jsx jsx */
 import { jsx } from "theme-ui"
 import { Component } from "react"
-import { mediaQueries } from "../../gatsby-plugin-theme-ui"
+import { mediaQueries } from "gatsby-design-tokens/dist/theme-gatsbyjs-org"
 import { renderCell } from "./evaluation-cell"
 import SectionTitle from "./evaluation-table-section-title"
 import SectionHeaderTop from "./evaluation-table-section-header-top"
@@ -53,8 +53,8 @@ class EvaluationTable extends Component {
                     [].concat([
                       <SectionHeaderBottom
                         options={options}
-                        display={row.node.Subcategory}
-                        category={row.node.Subcategory}
+                        display={row.Subcategory}
+                        category={row.Subcategory}
                         key={`section-header-${s}-bottom-${i}`}
                       />,
                       // table row with the name of the feature and corresponding scores
@@ -81,7 +81,7 @@ class EvaluationTable extends Component {
                             }}
                             id={
                               j === 0
-                                ? row.node.Feature.toLowerCase()
+                                ? row.Feature.toLowerCase()
                                     .split(` `)
                                     .join(`-`)
                                 : undefined
@@ -92,7 +92,7 @@ class EvaluationTable extends Component {
                               })
                             }}
                           >
-                            {renderCell(row.node[nodeProperty], j)}
+                            {renderCell(row[nodeProperty], j)}
                           </td>
                         ))}
                       </tr>,
@@ -117,7 +117,7 @@ class EvaluationTable extends Component {
                           {
                             <span
                               dangerouslySetInnerHTML={{
-                                __html: row.node.Description,
+                                __html: row.Description,
                               }}
                             />
                           }