+++ www/src/pages/index.js
@@ -5,7 +5,6 @@ import { graphql } from "gatsby"
 import { Helmet } from "react-helmet"
 import ArrowForwardIcon from "react-icons/lib/md/arrow-forward"
 
-import Layout from "../components/layout"
 import Container from "../components/container"
 import MastheadContent from "../components/masthead"
 import Diagram from "../components/diagram"
@@ -16,7 +15,6 @@ import HomepageFeatures from "../components/homepage/homepage-features"
 import HomepageEcosystem from "../components/homepage/homepage-ecosystem"
 import HomepageBlog from "../components/homepage/homepage-blog"
 import HomepageNewsletter from "../components/homepage/homepage-newsletter"
-import HomepageSection from "../components/homepage/homepage-section"
 import FooterLinks from "../components/shared/footer-links"
 import {
   setupScrollersObserver,
@@ -43,22 +41,20 @@ class IndexRoute extends React.Component {
   render() {
     const {
       data: {
-        allMdx: { edges: postsData },
-        allStartersYaml: { edges: startersData },
-        allNpmPackage: { edges: pluginsData },
+        allMdx: { nodes: postsData },
+        allStartersYaml: { nodes: startersData },
+        allNpmPackage: { nodes: pluginsData },
       },
     } = this.props
 
     const starters = startersData.map(item => {
       const {
-        node: {
-          fields: {
-            starterShowcase: { slug, name, description, stars },
-          },
-          childScreenshot: {
-            screenshotFile: {
-              childImageSharp: { fixed: thumbnail },
-            },
+        fields: {
+          starterShowcase: { slug, name, description, stars },
+        },
+        childScreenshot: {
+          screenshotFile: {
+            childImageSharp: { fixed: thumbnail },
           },
         },
       } = item
@@ -74,9 +70,9 @@ class IndexRoute extends React.Component {
     })
 
     const plugins = pluginsData.map(item => {
-      item.node.type = `Plugin`
+      item.type = `Plugin`
 
-      return item.node
+      return item
     })
 
     const ecosystemFeaturedItems = this.combineEcosystemFeaturedItems({
@@ -84,10 +80,8 @@ class IndexRoute extends React.Component {
       starters,
     })
 
-    const posts = postsData.map(item => item.node)
-
     return (
-      <Layout location={this.props.location}>
+      <>
         <Helmet>
           <meta
             name="Description"
@@ -138,19 +132,12 @@ class IndexRoute extends React.Component {
 
           <HomepageEcosystem featuredItems={ecosystemFeaturedItems} />
 
-          <HomepageBlog posts={posts} />
+          <HomepageBlog posts={postsData} />
 
           <HomepageNewsletter />
-
-          <HomepageSection
-            css={{
-              paddingTop: `0 !important`,
-              paddingBottom: `0 !important`,
-            }}
-          />
         </main>
         <FooterLinks />
-      </Layout>
+      </>
     )
   }
 }
@@ -158,10 +145,7 @@ class IndexRoute extends React.Component {
 export default IndexRoute
 
 export const pageQuery = graphql`
-  query IndexRouteQuery(
-    $featuredStarters: [String]!
-    $featuredPlugins: [String]!
-  ) {
+  query IndexRouteQuery {
     file(relativePath: { eq: "gatsby-explanation.png" }) {
       childImageSharp {
         fluid(maxWidth: 870) {
@@ -180,51 +164,42 @@ export const pageQuery = graphql`
         fields: { released: { eq: true } }
       }
     ) {
-      edges {
-        node {
-          ...HomepageBlogPostData
-        }
+      nodes {
+        ...HomepageBlogPostData
       }
     }
     allStartersYaml(
       filter: {
-        fields: {
-          starterShowcase: { slug: { in: $featuredStarters } }
-          hasScreenshot: { eq: true }
-        }
+        fields: { featured: { eq: true }, hasScreenshot: { eq: true } }
       }
       sort: { order: DESC, fields: [fields___starterShowcase___stars] }
     ) {
-      edges {
-        node {
-          fields {
-            starterShowcase {
-              slug
-              description
-              stars
-              name
-            }
+      nodes {
+        fields {
+          starterShowcase {
+            slug
+            description
+            stars
+            name
           }
-          childScreenshot {
-            screenshotFile {
-              childImageSharp {
-                fixed(width: 64, height: 64) {
-                  ...GatsbyImageSharpFixed_noBase64
-                }
+        }
+        childScreenshot {
+          screenshotFile {
+            childImageSharp {
+              fixed(width: 64, height: 64) {
+                ...GatsbyImageSharpFixed_noBase64
               }
             }
           }
         }
       }
     }
-    allNpmPackage(filter: { name: { in: $featuredPlugins } }) {
-      edges {
-        node {
-          slug
-          name
-          description
-          humanDownloadsLast30Days
-        }
+    allNpmPackage(filter: { fields: { featured: { eq: true } } }) {
+      nodes {
+        slug
+        name
+        description
+        humanDownloadsLast30Days
       }
     }
   }