+++ www/src/templates/tags.js
@@ -9,7 +9,7 @@ import TiArrowRight from "react-icons/lib/ti/arrow-right"
 import BlogPostPreviewItem from "../components/blog-post-preview-item"
 import Button from "../components/button"
 import Container from "../components/container"
-import Layout from "../components/layout"
+import FooterLinks from "../components/shared/footer-links"
 import { TAGS_AND_DOCS } from "../data/tags-docs"
 
 // Select first tag with whitespace instead of hyphens for
@@ -24,16 +24,16 @@ const preferSpacedTag = tags => {
   return tags[0]
 }
 
-const Tags = ({ pageContext, data, location }) => {
+const Tags = ({ pageContext, data }) => {
   const { tags } = pageContext
-  const { edges, totalCount } = data.allMdx
+  const { nodes, totalCount } = data.allMdx
   const tagHeader = `${totalCount} post${
     totalCount === 1 ? `` : `s`
   } tagged with "${preferSpacedTag(tags)}"`
   const doc = TAGS_AND_DOCS.get(tags[0])
 
   return (
-    <Layout location={location}>
+    <Container>
       <Helmet>
         <title>{`${preferSpacedTag(tags)} Tag`}</title>
         <meta
@@ -43,37 +43,36 @@ const Tags = ({ pageContext, data, location }) => {
           )}`}
         />
       </Helmet>
-      <Container>
-        <h1>{tagHeader}</h1>
-        <Button
-          variant="small"
-          key="blog-post-view-all-tags-button"
-          to="/blog/tags"
-        >
-          View all tags <TagsIcon />
-        </Button>
-        {doc ? (
-          <React.Fragment>
-            <span css={{ margin: 5 }} />
-            <Button
-              variant="small"
-              secondary
-              key={`view-tag-docs-button`}
-              to={doc}
-            >
-              Read the documentation <TiArrowRight />
-            </Button>
-          </React.Fragment>
-        ) : null}
-        {edges.map(({ node }) => (
-          <BlogPostPreviewItem
-            post={node}
-            key={node.fields.slug}
-            sx={{ my: 9 }}
-          />
-        ))}
-      </Container>
-    </Layout>
+      <h1>{tagHeader}</h1>
+      <Button
+        variant="small"
+        key="blog-post-view-all-tags-button"
+        to="/blog/tags"
+      >
+        View all tags <TagsIcon />
+      </Button>
+      {doc ? (
+        <React.Fragment>
+          <span css={{ margin: 5 }} />
+          <Button
+            variant="small"
+            secondary
+            key={`view-tag-docs-button`}
+            to={doc}
+          >
+            Read the documentation <TiArrowRight />
+          </Button>
+        </React.Fragment>
+      ) : null}
+      {nodes.map(node => (
+        <BlogPostPreviewItem
+          post={node}
+          key={node.fields.slug}
+          sx={{ my: 9 }}
+        />
+      ))}
+      <FooterLinks />
+    </Container>
   )
 }
 
@@ -91,10 +90,8 @@ export const pageQuery = graphql`
       }
     ) {
       totalCount
-      edges {
-        node {
-          ...BlogPostPreview_item
-        }
+      nodes {
+        ...BlogPostPreview_item
       }
     }
   }